configs: remote: add config for double boot strategy
Add config for DoubleBootStrategy. It uses two shell drivers, so this
config uses 'list' notation.

Note that in this example config NetworkService resource is specified
separately from RemotePlace. This is possible if DUT is accessible from
test machine. For instance consider the following scenario:
- DUT boots to 'setup' mode and becomes accessible from test machine

The more isolated testlab configuration may not allow to access DUT
from test machine. In this case NetworkService will be the part of
the RemotePlace. Though exporter will require additional tools and
configuration including 'socat' and 'labgrid-bound-connect', see:
- https://labgrid.readthedocs.io/en/v25.0/configuration.html#networkservice

Signed-off-by: Sergey Matyukevich <geomatsi@gmail.com>