modbus: connect to server only when needed
Currently connection to Modbus TCP server is opened on activation and
closed on deactivation. That means that connection is kept alive during
the time of the test session. This does not work well if multi-channel
relay board supports only limited number of opened connection. For
instance, this is the case for DTWonder relay board: it supports only
two simultaneous open connections.

In fact no need to keep connection to Modbus TCP opened all the time.
This change modifies ModbusTCPPowerDriver to open connection only for
each requested power operation and close right after its completion.

Signed-off-by: Sergey Matyukevich <geomatsi@gmail.com>