configs: remote: fix bananapi canmv k230d zero
The board’s UART MCU 5V power rail is supplied only from the USB Type-C
connector. When the board is powered via the expansion connector, the
UART is therefore unavailable.

To make the board usable in my lab setup, I soldered the UART MCU 5V
rail to the board’s VDD_5V rail. However, in this configuration the
UART becomes available in Labgrid only after a delay, once the
resource is detected.

To handle this, add the 'late console' quirk to the UbootNet strategy
for this board.

Signed-off-by: Sergey Matyukevich <geomatsi@gmail.com>